+++ src/TimeSeriesPanel.tsx
@@ -18,7 +18,6 @@ import { FrameSettingsEditor } from 'plugins/frameSettings/FrameSettingsEditor';
 import { FieldSettings } from 'app/types/frameSettings';
 import { Options } from './panelcfg.gen';
 import { ExemplarsPlugin, getVisibleLabels } from './plugins/ExemplarsPlugin';
-import { OutsideRangePlugin } from './plugins/OutsideRangePlugin';
 import { ThresholdControlsPlugin } from './plugins/ThresholdControlsPlugin';
 import { TimescaleEditor } from './plugins/timescales/TimescaleEditor';
 import { TimescaleItem } from './plugins/timescales/TimescaleEditorForm';
@@ -30,6 +29,7 @@ import { TimeSeriesTooltip } from './TimeSeriesTooltip';
 import { AnnotationsPlugin2 } from './plugins/AnnotationsPlugin2';
 import { PinnedTooltip } from 'app/core/components/PinnedTooltip/PinnedTooltip';
 import { PinnedPoint } from 'app/types';
+import { OutsideConfigPlugins } from 'plugins/OutsideConfigPlugins';
 
 interface TimeSeriesPanelProps extends PanelProps<Options> {}
 
@@ -63,6 +63,9 @@ export const TimeSeriesPanel = ({
     eventBus,
   } = usePanelContext();
 
+  /**
+   * State
+   */
   const [isAddingTimescale, setAddingTimescale] = useState(false);
   const [fieldSettings, setFieldSettings] = useState<FieldSettings[]>([]);
   const [showFrameSettings, setShowFrameSettings] = useState(false);
@@ -117,7 +120,17 @@ export const TimeSeriesPanel = ({
 
   const timezones = useMemo(() => getTimezones(options.timezone, timeZone), [options.timezone, timeZone]);
 
+  /**
+   * Timescales Frame
+   * scales per well
+   */
   const [timescalesFrame, setTimescalesFrame] = useState<DataFrame | null>(null);
+
+  /**
+   * Timescales Frame
+   * global scales settings well eq. "GLOBAL_SCALES"
+   */
+  const [globalTimescalesFrame, setGlobalTimescalesFrame] = useState<DataFrame | null>(null);
 
   const mappings = fieldConfig.defaults.mappings;
   let scales: string[] = [];
@@ -144,50 +157,114 @@ export const TimeSeriesPanel = ({
     well = Array.isArray(wellVariable.current.value) ? wellVariable.current.value[0] : wellVariable.current.value;
   }
 
-  const getTimescales = useCallback(async () => {
-    const user = config.bootData.user;
-    const userId = user?.id;
-    const rawSql = `select min, max, metric from scales where user_id=${userId} and well='${well}';`;
-    const target = data.request?.targets[0];
-    const datasourceId = target?.datasource?.uid;
-    const refId = target?.refId;
+  /**
+   * Get timescales
+   * scales per well
+   */
+  const getTimescales = useCallback(
+    async (wellValue: string | null) => {
+      const user = config.bootData.user;
+      const userId = user?.id;
+      const dashboardId = data.request?.dashboardUID;
 
-    if (refId) {
-      const response = await getBackendSrv().post('/api/ds/query', {
-        debug: true,
-        from: 'now-1h',
-        publicDashboardAccessToken: 'string',
-        queries: [
-          {
-            datasource: {
-              uid: datasourceId,
+      let rawSql = '';
+      if (!wellValue) {
+        /**
+         * Get
+         * global scales, well eq. "GLOBAL_SCALES"
+         */
+        rawSql = `select min, max, metric from scales where user_id=${userId} and dashboard_id='${dashboardId}' and well='GLOBAL_SCALES';`;
+      } else {
+        /**
+         * Get
+         * scales, based on well
+         */
+        rawSql = `select min, max, metric from scales where user_id=${userId} and dashboard_id='${dashboardId}' and well='${well}';`;
+      }
+
+      const target = data.request?.targets[0];
+      const datasourceId = target?.datasource?.uid;
+      const refId = target?.refId;
+
+      if (refId) {
+        const response = await getBackendSrv().post('/api/ds/query', {
+          debug: true,
+          from: 'now-1h',
+          publicDashboardAccessToken: 'string',
+          queries: [
+            {
+              datasource: {
+                uid: datasourceId,
+              },
+              format: 'table',
+              intervalMs: 86400000,
+              maxDataPoints: 1092,
+              rawSql,
+              refId,
             },
-            format: 'table',
-            intervalMs: 86400000,
-            maxDataPoints: 1092,
-            rawSql,
-            refId,
-          },
-        ],
-        to: 'now',
-      });
+          ],
+          to: 'now',
+        });
 
-      setTimescalesFrame(toDataFrame(response.results?.[refId]?.frames[0]));
-      return;
-    }
+        if (!wellValue) {
+          /**
+           * Global scales, well eq. "GLOBAL_SCALES"
+           */
+          setGlobalTimescalesFrame(toDataFrame(response.results?.[refId]?.frames[0]));
+        } else {
+          /**
+           * Scales, based on well
+           */
+          setTimescalesFrame(toDataFrame(response.results?.[refId]?.frames[0]));
+        }
 
-    setTimescalesFrame(null);
-  }, [data.request?.targets, well]);
+        return;
+      }
+
+      if (!wellValue) {
+        /**
+         * Global scales, well eq. "GLOBAL_SCALES"
+         */
+        setGlobalTimescalesFrame(null);
+      } else {
+        /**
+         * Scales, based on well
+         */
+        setTimescalesFrame(null);
+      }
+    },
+    [data.request?.targets, well, data.request?.dashboardUID]
+  );
 
   const onUpsertTimescale = useCallback(
-    async (formData: TimescaleItem) => {
+    async (formData: TimescaleItem, isGlobal: boolean) => {
       const { min, max, description, scale, auto } = formData;
       const user = config.bootData.user;
       const userId = user?.id;
+      const dashboardId = data.request?.dashboardUID;
       const sanitizedDescription = description.replace(/\"|\'/g, '');
-      const rawSql = `insert into scales values ('${well}', ${userId}, '${scale}', ${auto ? null : min}, ${
-        auto ? null : max
-      }, '${sanitizedDescription}') on conflict (well, user_id, metric) do update set min = excluded.min, max = excluded.max;`;
+      let rawSql = '';
+
+      if (!isGlobal) {
+        /**
+         * Scales, based on well
+         */
+        rawSql = `insert into scales (well, user_id, dashboard_id, metric, min, max, description) \
+        values ('${well}', ${userId}, '${dashboardId}', '${scale}', ${auto ? null : min}, ${
+          auto ? null : max
+        }, '${sanitizedDescription}') on conflict (well, user_id, dashboard_id, metric) do \
+        update set min = excluded.min, max = excluded.max;`;
+      } else {
+        /**
+         * Global scales, well eq. "GLOBAL_SCALES"
+         */
+        rawSql = `insert into scales (well, user_id, dashboard_id, metric, min, max, description) \
+        values ('GLOBAL_SCALES', ${userId}, '${dashboardId}', '${scale}', ${auto ? null : min}, ${
+          auto ? null : max
+        }, '${sanitizedDescription}') on conflict (well, user_id, dashboard_id, metric) do \
+        update set min = excluded.min, max = excluded.max;`;
+      }
+
       const target = data.request?.targets[0];
       const datasourceId = target?.datasource?.uid;
       const refId = target?.refId;
@@ -211,12 +288,12 @@ export const TimeSeriesPanel = ({
         to: 'now',
       });
     },
-    [data.request?.targets, well]
+    [data.request?.dashboardUID, data.request?.targets, well]
   );
 
   const onUpsertTimescales = useCallback(
-    async (timescales: TimescaleItem[]) => {
-      await Promise.all(timescales.map((timescale) => onUpsertTimescale(timescale)));
+    async (timescales: TimescaleItem[], isGlobal: boolean) => {
+      await Promise.all(timescales.map((timescale) => onUpsertTimescale(timescale, isGlobal)));
 
       /**
        * Refresh Dashboard
@@ -226,9 +303,14 @@ export const TimeSeriesPanel = ({
       /**
        * Refresh timescales
        */
-      await getTimescales();
+      await getTimescales(well);
+
+      /**
+       * Global scales, well eq. "GLOBAL_SCALES"
+       */
+      await getTimescales(null);
     },
-    [getTimescales, onUpsertTimescale, dashboardRefresh]
+    [dashboardRefresh, getTimescales, well, onUpsertTimescale]
   );
 
   const suggestions = useMemo(() => {
@@ -422,7 +504,8 @@ export const TimeSeriesPanel = ({
                                   top: u.rect.top + (u.cursor.top ?? 0),
                                 });
                                 setAddingTimescale(true);
-                                getTimescales();
+                                getTimescales(well);
+                                getTimescales(null);
                                 dismiss();
                               }}
                             >
@@ -481,7 +564,23 @@ export const TimeSeriesPanel = ({
                     options={options}
                     variable={wellVariable}
                   />
-                  <OutsideRangePlugin config={uplotConfig} onChangeTimeRange={onChangeTimeRange} />
+                  <OutsideConfigPlugins
+                    config={uplotConfig}
+                    onChangeTimeRange={onChangeTimeRange}
+                    frames={frames}
+                    customScalesHandler={() => {
+                      if (panelRoot.current) {
+                        /**
+                         * setTriggerCoords
+                         */
+                        const { right, bottom } = panelRoot.current?.getBoundingClientRect();
+                        setTriggerCoords({ left: right / 2, top: bottom / 2 });
+                      }
+                      setAddingTimescale(true);
+                      getTimescales(well);
+                      getTimescales(null);
+                    }}
+                  />
                   {data.annotations && (
                     <ExemplarsPlugin
                       visibleSeries={getVisibleLabels(uplotConfig, frames)}
@@ -510,6 +609,7 @@ export const TimeSeriesPanel = ({
                     top: triggerCoords?.top,
                   }}
                   timescalesFrame={timescalesFrame}
+                  globalTimescalesFrame={globalTimescalesFrame}
                 />
               )}
               {showFrameSettings && (